requirements.txt
bumpversion==0.5.3
flask
gevent
gunicorn
jsonschema
python-dateutil
requests>=2.20.0
Sphinx
watchdog==0.8.3
#wheel==0.23.0
psycopg2-binary
ipwhois==1.1.0